@@ -2,11 +2,10 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace CT.Core.Shared.ModelsViews
+namespace CT.Core.Shared.ModelsViews;
+
+public class EspecialidadeView
 {
-    public class EspecialidadeView
-    {
-        public int Id { get; set; }
-        public string Descricao { get; set; }
-    }
+    public int Id { get; set; }
+    public string Descricao { get; set; }
 }
