@@ -2,10 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace CT.Core.Shared.ModelsViews.Medico
+namespace CT.Core.Shared.ModelsViews.Medico;
+
+public class AlteraMedico : NovoMedico
 {
-    public class AlteraMedico : NovoMedico
-    {
-        public int Id { get; set; }
-    }
+    public int Id { get; set; }
 }
