@@ -2,16 +2,16 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace CT.Data.Configuration
+namespace CT.Data.Configuration;
+
+public class EnderecoConfiguration : IEntityTypeConfiguration<Endereco>
 {
-    public class EnderecoConfiguration : IEntityTypeConfiguration<Endereco>
+    public void Configure(EntityTypeBuilder<Endereco> builder)
     {
-        public void Configure(EntityTypeBuilder<Endereco> builder)
-        {
-            builder.HasKey(p => p.ClienteId);
-        }
+        builder.HasKey(e => e.ClienteId);
+        builder.Property(e => e.Estado).HasConversion(
+            e => e.ToString(),
+            e => (Estado)Enum.Parse(typeof(Estado), e));
     }
 }
