@@ -1,17 +1,13 @@
 ﻿using CT.Core.Domain;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Text;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace CT.Data.Configuration
+namespace CT.Data.Configuration;
+
+public class TelefoneConfiguration : IEntityTypeConfiguration<Telefone>
 {
-    public class TelefoneConfiguration : IEntityTypeConfiguration<Telefone>
+    public void Configure(EntityTypeBuilder<Telefone> builder)
     {
-        public void Configure(Microsoft.EntityFrameworkCore.Metadata.Builders.EntityTypeBuilder<Telefone> builder)
-        {
-            //builder.HasOne(p => p.Cliente).WithMany(p => p.Telefones).IsRequired().OnDelete(DeleteBehavior.Cascade);
-            builder.HasKey(p => new { p.ClienteId, p.Numero });//Chave composta
-        }
+        builder.HasKey(p => new { p.ClienteId, p.Numero });
     }
 }
