@@ -1,66 +1,79 @@
 ﻿using CT.Core.Domain;
 using CT.Data.Context;
-using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
-using System.Linq;
 using Microsoft.EntityFrameworkCore;
-using CT.Manager.Interfaces;
+using CT.Manager.Interfaces.Repositories;
 
-namespace CT.Data.Repository
+namespace CT.Data.Repository;
+
+public class ClienteRepository : IClienteRepository
 {
-    public class ClienteRepository : IClienteRepository
+    private readonly AppDbContext _context;
+
+    public ClienteRepository(AppDbContext context)
     {
-        private readonly AppDbContext _context;
+        _context = context;
+    }
 
-        public ClienteRepository(AppDbContext context)
+    public async Task<IEnumerable<Cliente>> GetClientesAsync()
+    {
+        return await _context.Clientes
+            .Include(p => p.Endereco)
+            .Include(p => p.Telefones)
+            .AsNoTracking().ToListAsync();
+    }
+
+    public async Task<Cliente> GetClienteAsync(int id)
+    {
+        return await _context.Clientes
+            .Include(p => p.Endereco)
+            .Include(p => p.Telefones)
+            .SingleOrDefaultAsync(p => p.Id == id);
+    }
+
+    public async Task<Cliente> InsertClienteAsync(Cliente cliente)
+    {
+        await _context.Clientes.AddAsync(cliente);
+        await _context.SaveChangesAsync();
+        return cliente;
+    }
+
+    public async Task<Cliente> UpdateClienteAsync(Cliente cliente)
+    {
+        var clienteConsultado = await _context.Clientes
+                                             .Include(p => p.Endereco)
+                                             .Include(p => p.Telefones)
+                                             .FirstOrDefaultAsync(p => p.Id == cliente.Id);
+        if (clienteConsultado == null)
         {
-            _context = context;
+            return null;
         }
+        _context.Entry(clienteConsultado).CurrentValues.SetValues(cliente);
+        clienteConsultado.Endereco = cliente.Endereco;
+        UpdateClienteTelefones(cliente, clienteConsultado);
+        await _context.SaveChangesAsync();
+        return clienteConsultado;
+    }
 
-        public async Task<IEnumerable<Cliente>> GetClientesAsync()
+    private static void UpdateClienteTelefones(Cliente cliente, Cliente clienteConsultado)
+    {
+        clienteConsultado.Telefones.Clear();
+        foreach (var telefone in cliente.Telefones)
         {
-            return await _context.Clientes
-                .Include(c => c.Endereco)
-                .Include(c => c.Telefones)
-                .AsNoTracking().ToListAsync();
+            clienteConsultado.Telefones.Add(telefone);
         }
+    }
 
-        public async Task<Cliente> GetClienteByIdAsync(int id)
+    public async Task<Cliente> DeleteClienteAsync(int id)
+    {
+        var clienteConsultado = await _context.Clientes.FindAsync(id);
+        if (clienteConsultado == null)
         {
-            return await _context.Clientes
-                .Include(c => c.Endereco)
-                .Include(c => c.Telefones)
-                .SingleOrDefaultAsync(c => c.Id == id);
+            return null;
         }
-
-        public async Task<Cliente> InsertClienteAsync(Cliente cliente)
-        {
-            await _context.Clientes.AddAsync(cliente);
-            await _context.SaveChangesAsync();
-            return cliente;
-        }
-
-        public async Task<Cliente> UpdateClienteByIdAsync(Cliente cliente)
-        {
-            var clienteConsultado = await _context.Clientes.FindAsync(cliente.Id);
-            if (clienteConsultado == null)
-            {
-                return null;
-            }
-
-            //Atribuindo valores
-            _context.Entry(clienteConsultado).CurrentValues.SetValues(cliente);
-            await _context.SaveChangesAsync();
-            return clienteConsultado;
-        }
-
-        public async Task DeleteClienteByIdAsync(int id)
-        {
-            var clienteConsultado = await _context.Clientes.FindAsync(id);
-            _context.Clientes.Remove(clienteConsultado);
-            await _context.SaveChangesAsync();
-        }
+        var clienteRemovido = _context.Clientes.Remove(clienteConsultado);
+        await _context.SaveChangesAsync();
+        return clienteRemovido.Entity;
     }
 }
