@@ -1,25 +1,21 @@
 ﻿using CT.Data.Context;
-using CT.Manager.Interfaces;
+using CT.Manager.Interfaces.Repositories;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace CT.Data.Repository
+namespace CT.Data.Repository;
+
+public class EspecialidadeRepository : IEspecialidadeRepository
 {
-    public class EspecialidadeRepository : IEspecialidadeRepository
+    private readonly AppDbContext _context;
+
+    public EspecialidadeRepository(AppDbContext context)
     {
-        private readonly AppDbContext _context;
+        _context = context;
+    }
 
-        public EspecialidadeRepository(AppDbContext context)
-        {
-            _context = context;
-        }
-
-        public async Task<bool> ExisteAsync(int id)
-        {
-            return await _context.Especialidades.AnyAsync(p => p.Id == id);
-        }
+    public async Task<bool> ExisteAsync(int id)
+    {
+        return await _context.Especialidades.AnyAsync(p => p.Id == id);
     }
 }
