@@ -5,25 +5,21 @@ using CT.Core.Shared.ModelsViews.Cliente;
 using CT.Core.Shared.ModelsViews.Endereco;
 using CT.Core.Shared.ModelsViews.Telefone;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace CT.Manager.Mappings
+namespace CT.Manager.Mappings;
+
+public class NovoClienteMappingProfile : Profile
 {
-    public class NovoClienteMappingProfile : Profile
+    public NovoClienteMappingProfile()
     {
-        public NovoClienteMappingProfile()
-        {
-            CreateMap<NovoCliente, Cliente>()
-                .ForMember(d => d.Criacao, o => o.MapFrom(c => DateTime.Now))
-                .ForMember(d => d.DataNascimento, o => o.MapFrom(c => c.DataNascimento.Date))
-                //.ReverseMap()
-                ;
-            CreateMap<NovoEndereco, Endereco>();
-            CreateMap<NovoTelefone, Telefone>();
-            CreateMap<Cliente, ClienteView>();
-            CreateMap<Endereco, EnderecoView>();
-            CreateMap<Telefone, TelefoneView>();
-        }
+        CreateMap<NovoCliente, Cliente>()
+            .ForMember(d => d.Criacao, o => o.MapFrom(_ => DateTime.Now))
+            .ForMember(d => d.DataNascimento, o => o.MapFrom(x => x.DataNascimento.Date));
+
+        CreateMap<NovoEndereco, Endereco>();
+        CreateMap<NovoTelefone, Telefone>();
+        CreateMap<Cliente, ClienteView>();
+        CreateMap<Endereco, EnderecoView>();
+        CreateMap<Telefone, TelefoneView>();
     }
 }
