@@ -2,22 +2,18 @@
 using CT.Core.Domain;
 using CT.Core.Shared.ModelsViews;
 using CT.Core.Shared.ModelsViews.Medico;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace CT.Manager.Mappings
+namespace CT.Manager.Mappings;
+
+public class NovoMedicoMappingProfile : Profile
 {
-    public class NovoMedicoMappingProfile : Profile
+    public NovoMedicoMappingProfile()
     {
-        public NovoMedicoMappingProfile()
-        {
-            CreateMap<NovoMedico, Medico>();
-            CreateMap<Medico, MedicoView>();
-            CreateMap<Especialidade, ReferenciaEspecialidade>().ReverseMap();
-            CreateMap<Especialidade, EspecialidadeView>().ReverseMap();
-            CreateMap<Especialidade, NovaEspecialidade>().ReverseMap();
-            CreateMap<AlteraMedico, Medico>().ReverseMap();
-        }
+        CreateMap<NovoMedico, Medico>();
+        CreateMap<Medico, MedicoView>();
+        CreateMap<Especialidade, ReferenciaEspecialidade>().ReverseMap();
+        CreateMap<Especialidade, EspecialidadeView>().ReverseMap();
+        CreateMap<Especialidade, NovaEspecialidade>().ReverseMap();
+        CreateMap<AlteraMedico, Medico>().ReverseMap();
     }
 }
