@@ -1,47 +1,38 @@
-﻿using CT.Core.Domain;
-using CT.Core.Shared.ModelsViews;
+﻿using CT.Core.Shared.ModelsViews;
 using FluentValidation;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace CT.Manager.Validator
+namespace CT.Manager.Validator;
+
+public class NovoClienteValidator : AbstractValidator<NovoCliente>
 {
-    public class NovoClienteValidator : AbstractValidator<NovoCliente>
+    public NovoClienteValidator()
     {
-        public NovoClienteValidator()
-        {
-            RuleFor(x => x.Nome)
-                .NotNull()
-                .NotEmpty()
-                .MinimumLength(10)
-                .MaximumLength(150);
+        RuleFor(x => x.Nome)
+            .NotNull()
+            .NotEmpty()
+            .MinimumLength(10)
+            .MaximumLength(150);
 
-            RuleFor(x => x.DataNascimento)
-                .NotNull()
-                .NotEmpty()
-                .LessThan(DateTime.Now)
-                .GreaterThan(DateTime.Now
-                .AddYears(-130));
+        RuleFor(x => x.DataNascimento)
+            .NotNull()
+            .NotEmpty()
+            .LessThan(DateTime.Now)
+            .GreaterThan(DateTime.Now
+            .AddYears(-130));
 
-            RuleFor(x => x.Documento)
-                .NotNull().NotEmpty()
-                .MinimumLength(4)
-                .MaximumLength(14);
+        RuleFor(x => x.Documento)
+            .NotNull().NotEmpty()
+            .MinimumLength(4)
+            .MaximumLength(14);
 
-            RuleFor(x => x.Telefones)
-                .NotNull()
-                .NotEmpty();
+        RuleFor(x => x.Telefones)
+            .NotNull()
+            .NotEmpty();
 
-            RuleFor(x => x.Sexo).NotNull().NotEmpty().Must(IsMorF).WithMessage("Sexo precisa ser M ou F");
+        RuleFor(x => x.Sexo).NotNull();
 
-            RuleFor(x => x.Endereco)
-                .SetValidator(new NovoEnderecoValidator());
-        }
-
-        private bool IsMorF(char sexo)
-        {
-            return sexo == 'M' || sexo == 'F';
-        }
+        RuleFor(x => x.Endereco)
+            .SetValidator(new NovoEnderecoValidator());
     }
 }
