@@ -1,19 +1,17 @@
 ﻿using CT.Manager.Mappings;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace CT.WebApi.Configuration
+namespace CT.WebApi.Configuration;
+
+public static class AutoMapperConfig
 {
-    public static class AutoMapperConfig
+    public static void AddAutoMapperConfiguration(this IServiceCollection services)
     {
-        public static void AddAutoMapperConfiguration(this IServiceCollection services)
-        {
-            services.AddAutoMapper(
-                typeof(NovoClienteMappingProfile),
-                typeof(AlteraClienteMappingProfile),
-                typeof(NovoMedicoMappingProfile)
-            //,typeof(UsuarioMappingProfile)
-                );
-        }
+        services.AddAutoMapper(
+            typeof(NovoClienteMappingProfile),
+            typeof(AlteraClienteMappingProfile),
+            typeof(NovoMedicoMappingProfile)
+            );
     }
 }
 
