@@ -1,4 +1,4 @@
-using Microsoft.AspNetCore.Hosting;
+﻿using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -9,57 +9,53 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace CT.WebApi
+namespace CT.WebApi;
+
+public class Program
 {
-    public class Program
+    public static void Main(string[] args)
     {
-        public static void Main(string[] args)
+        IConfigurationRoot configuration = GetConfiguration();
+
+        ConfiguraLog(configuration);
+
+        try
         {
-            IConfigurationRoot configuration = GetConfiguration();
-
-            ConfigureLog(configuration);
-
-            try
-            {
-                Log.Information("Iniciando o WebApi");
-                CreateHostBuilder(args).Build().Run();
-            }
-            catch (Exception ex)
-            {
-                Log.Fatal(ex, "Erro castrat�fico.");
-                throw;
-            }
-            finally
-            {
-                Log.CloseAndFlush();
-            }
+            Log.Information("Iniciando o WebApi");
+            CreateHostBuilder(args).Build().Run();
         }
-
-        private static void ConfigureLog(IConfigurationRoot configuration)
+        catch (Exception ex)
         {
-            Log.Logger = new LoggerConfiguration()
-                            .ReadFrom.Configuration(configuration)
-                            .CreateLogger();
+            Log.Fatal(ex, "Erro catastr�fico.");
+            throw;
         }
-
-        private static IConfigurationRoot GetConfiguration()
+        finally
         {
-            string ambiente = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
-
-            var configuration = new ConfigurationBuilder()
-                .SetBasePath(Directory.GetCurrentDirectory())
-                .AddJsonFile("appsettings.json")
-                .AddJsonFile($"appsettings.{ambiente}.json", optional: true)
-                .Build();
-            return configuration;
+            Log.CloseAndFlush();
         }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .UseSerilog()
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
     }
+
+    private static void ConfiguraLog(IConfigurationRoot configuration)
+    {
+        Log.Logger = new LoggerConfiguration()
+            .ReadFrom.Configuration(configuration)
+            .CreateLogger();
+    }
+
+    private static IConfigurationRoot GetConfiguration()
+    {
+        string ambiente = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
+
+        var configuration = new ConfigurationBuilder()
+            .SetBasePath(Directory.GetCurrentDirectory())
+            .AddJsonFile("appsettings.json")
+            .AddJsonFile($"appsettings.{ambiente}.json", optional: true)
+            .Build();
+        return configuration;
+    }
+
+    public static IHostBuilder CreateHostBuilder(string[] args) =>
+        Host.CreateDefaultBuilder(args)
+            .UseSerilog()
+            .ConfigureWebHostDefaults(webBuilder => webBuilder.UseStartup<Startup>());
 }
