@@ -9,7 +9,7 @@ namespace CT.Core.Domain
         public int Id { get; set; }
         public string Nome { get; set; }
         public DateTime DataNascimento { get; set; }
-        public char Sexo { get; set; }
+        public Sexo Sexo { get; set; }
         public ICollection<Telefone> Telefones { get; set; }
         public string Documento { get; set; }
         public DateTime Criacao { get; set; }
