@@ -9,7 +9,7 @@ namespace CT.Core.Domain
     {
         public int ClienteId { get; set; }
         public int Cep { get; set; }
-        public string Estado { get; set; }
+        public Estado Estado { get; set; }
         public string Cidade { get; set; }
         public string Logradouro { get; set; }
         public string Numero { get; set; }
